# docs/swagger.yaml
definitions:
...
  CreateSpareVO:
    properties:
      code:
        description: 备件编号
        type: string
      flavor:
        description: 备件规格
        type: string
      material:
        description: 材质
        type: string
      name:
        description: 备件名称
        type: string
    required:
    - code
    type: object
...
  DeleteVO:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    type: object
...
  PageSpareOutVO:
    properties:
      code:
        description: 备件编号
        type: string
      flavor:
        description: 备件规格
        type: string
      gmtCreated:
        description: 创建时间
        type: string
      gmtUpdated:
        description: 修改时间
        type: string
      id:
        description: ID
        type: integer
      material:
        description: 材质
        type: string
      name:
        description: 备件名称
        type: string
    type: object
...
  SpareImportParsedVOS:
    properties:
      data:
        description: 导入数据
        items:
          $ref: '#/definitions/SpareImportVO'
        minItems: 1
        type: array
    type: object
  SpareImportVO:
    properties:
      code:
        description: 备件编号
        type: string
      flavor:
        description: 备件规格
        type: string
      material:
        description: 材质
        type: string
      name:
        description: 备件名称
        type: string
    required:
    - code
    type: object
...
  UpdateSpareVO:
    properties:
      code:
        description: 备件编号
        type: string
      flavor:
        description: 备件规格
        type: string
      id:
        description: ID
        type: integer
      material:
        description: 材质
        type: string
      name:
        description: 备件名称
        type: string
    required:
    - code
    - id
    type: object
...
info:
  contact: {}
  description: 模具管理API
  title: 模具管理API
  version: "1.0"
paths:
...
  /spare:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/DeleteVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 删除备件
      tags:
      - 备件
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/CreateSpareVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 添加备件
      tags:
      - 备件
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/UpdateSpareVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 更新备件
      tags:
      - 备件
  /spare/excel:
    get:
      consumes:
      - application/json
      parameters:
      - description: 指定ID导出备件
        in: query
        name: ids
        type: array
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 导出
      tags:
      - 备件
    post:
      consumes:
      - application/json
      parameters:
      - description: SpareImportParsedVOS
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SpareImportParsedVOS'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 导入解析后的Excel内容
      tags:
      - 备件
...
  /spare/one:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PageSpareOutVO'
      summary: 用户备件查看
      tags:
      - 备件
...
  /spare/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 备件编号
        in: query
        name: code
        type: string
      - description: 备件编号和名称模糊查找，该参数有值时忽略其它条件参数
        in: query
        name: codeOrName
        type: string
      - description: 当前页
        in: query
        name: currentPage
        type: integer
      - description: 备件规格
        in: query
        name: flavor
        type: string
      - description: 材质
        in: query
        name: material
        type: string
      - description: 备件名称
        in: query
        name: name
        type: string
      - description: 每页条数
        in: query
        name: size
        type: integer
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PageSpareOutVO'
      summary: 用户备件分页
      tags:
      - 备件
  /spare/parse:
    post:
      consumes:
      - application/json
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 解析Excel文件内容
      tags:
      - 备件
...
swagger: "2.0"